refactor(ui): move HealthPageChrome into @repowise-dev/ui/health
The page-chrome that wraps the four /health/* dashboard pages used to
live in packages/web and pulled in next/link, which kept downstream
consumers of @repowise-dev/ui from reusing it.

- New framework-agnostic packages/ui/src/health/page-chrome.tsx accepts
  a renderLink render-prop (already the pattern used by HealthTabs) and
  forwards it down — no next dep in ui/.
- HealthTabs gains an optional basePath prop so callers with a different
  URL shape (snapshot-scoped, owner/name-scoped, etc.) can override the
  /repos/${repoId}/health prefix without forking the component.
- packages/web/src/components/health/health-page-chrome.tsx becomes a
  thin Next-Link binding around the shared base — the four /health/*
  pages keep their existing import path with no behavioral change.
